Change the file structure of dashboard page
